+++ lib/errors.js
@@ -1,5 +1,5 @@
 // add errors here
-var httpErrors = {
+var errors = {
   NotFoundError: {
     message: 'resource not found',
     status: '404'
@@ -15,8 +15,8 @@ var httpErrors = {
 };
 
 // initialize errors
-var errors = module.exports.errors = Object.keys(httpErrors).reduce(function (memo, err) {
-  var defaultErrorProps = httpErrors[err];
+module.exports = Object.keys(errors).reduce(function (memo, err) {
+  var defaultErrorProps = errors[err];
 
   // provide message override
   memo[err] = function (message) {
@@ -35,22 +35,3 @@ var errors = module.exports.errors = Object.keys(httpErrors).reduce(function (me
 
   return memo;
 }, {});
-
-function attachErrorsMdw(req, res, next) {
-  req.error = errors;
-  next();
-}
-module.exports.attachErrorsMdw = attachErrorsMdw;
-
-function errorHandlerMdw(err, req, res, next) {
-  if (!err) return next();
-
-  res.setHeader('content-type', 'application/json');
-  res.status(err.status);
-  res.json(err.body);
-
-  req.log.error({error: err.body}, 'an error occurred while servicing this request');
-  return next();
-}
-module.exports.errorHandlerMdw = errorHandlerMdw;
-
